relocate shareable karma code
