+++ karma/componentStub.js
